requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-cors-headers==3.1.0
django-redis==4.10.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
Pillow==6.1.0
public==2019.4.13
PyJWT==1.7.1
pytz==2019.2
redis==3.3.8
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
